requirements.txt
-i https://pypi.org/simple
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.29
smmap==5.0.0 ; python_version >= '3.6'
